Moved IcoSphere and Sphere to Geometry package.
